Face-Recognition-Attendance-Projects/requirements.txt
streamlit==1.9.0
av==9.2.0
streamlit_webrtc
certifi==2020.6.20
chardet==3.0.4
pyodbc==4.0.32
click==7.1.2
decorator==4.4.2
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.3
opencv-python==4.5.5.64
Pillow==9.1.1
proglog==0.1.9
requests==2.24.0
tqdm==4.51.0
urllib3==1.25.11
wincertstore==0.2

